Revert the menu list editing and update only the show components of agend item list instead of all of them.
